restructuring the repo
