+++ chemvae/sampled_rnn_tf.py
