--- chemvae_torch/datasets.py
@@ -1,16 +1,0 @@
-"""Python module to generate a Dataset object for this project.
-
-See ((dataset URL without parentheses goes here)) for more information.
-
-
-Uncomment the following code block
-to allow your dataset to be imported from your notebooks
-using:
-
-    from stability_vae.datasets import dataset
-"""
-
-# from mtx_pyclient.models import ModelsApi
-# api = ModelsApi(environment="prod")
-# dataset = api.get_dataset(dataset_id)  # change dataset_id to the actual integer!
-# print("Don't forget to call `dataset.checkout()`!")
