--- chemvae_torch/example.py
@@ -1,9 +1,0 @@
-"""Example module for stability_vae."""
-
-
-def hello_world() -> str:
-    """Hello command.
-
-    :returns: A string.
-    """
-    return "Hello, world!"
