+++ scripts/train_vae_torch.py
@@ -18,6 +18,7 @@ import argparse
 import os
 import yaml
 import chemvae_torch.mol_utils as mu
+from chemvae_torch.utils_torch import vectorize_data, hot_to_smiles, sigmoid_schedule
 
 # Define the models, loss functions, etc.
 # You need to define your PyTorch models and other necessary components here.
@@ -29,149 +30,6 @@ import chemvae_torch.mol_utils as mu
 # encoder = YourEncoderModel()
 # decoder = YourDecoderModel()
 # property_predictor = YourPropertyPredictorModel()
-
-def hot_to_smiles(hot_x, indices_chars):
-    
-    assert type(hot_x) == np.ndarray
-    assert len(hot_x.shape) == 3
-    
-    smiles = []
-    for i in range(hot_x.shape[0]):  # number of samples
-        temp_str = ""
-        for j in range(hot_x.shape[1]):  # length of smiles
-            index = np.argmax(hot_x[i, j, :])
-            temp_str += indices_chars[index]
-        smiles.append(temp_str)
-    return smiles
-
-def smiles_to_hot(smiles, params, canonize_smiles=True, check_smiles=False):
-
-    chars = yaml.safe_load(open(params["char_file"]))
-    char_indices = dict((c, i) for i, c in enumerate(chars))
-    indices_char = dict((i, c) for i, c in enumerate(chars))
-
-    if isinstance(smiles, str):
-        smiles = [smiles]
-
-    if canonize_smiles:
-        smiles = [mu.canon_smiles(s) for s in smiles]
-
-    if check_smiles:
-        smiles = mu.smiles_to_hot_filter(smiles, char_indices)
-
-    z = mu.smiles_to_hot(smiles, params["MAX_LEN"], params["PADDING"], char_indices, params["NCHARS"])
-    return z
-
-def vectorize_data(params):
-    # @out : Y_train /Y_test : each is list of datasets.
-    #        i.e. if reg_tasks only : Y_train_reg = Y_train[0]
-    #             if logit_tasks only : Y_train_logit = Y_train[0]
-    #             if both reg and logit_tasks : Y_train_reg = Y_train[0], Y_train_reg = 1
-    #             if no prop tasks : Y_train = []
-
-    MAX_LEN = params['MAX_LEN']
-
-    CHARS = yaml.safe_load(open(params['char_file']))
-    params['NCHARS'] = len(CHARS)
-    NCHARS = len(CHARS)
-    CHAR_INDICES = dict((c, i) for i, c in enumerate(CHARS))
-    #INDICES_CHAR = dict((i, c) for i, c in enumerate(CHARS))
-
-    ## Load data for properties
-    if params['do_prop_pred'] and ('data_file' in params):
-        if "data_normalization_out" in params:
-            normalize_out = params["data_normalization_out"]
-        else:
-            normalize_out = None
-
-        ################
-        if ("reg_prop_tasks" in params) and ("logit_prop_tasks" in params):
-            smiles, Y_reg, Y_logit = mu.load_smiles_and_data_df(params['data_file'], MAX_LEN,
-                    reg_tasks=params['reg_prop_tasks'], logit_tasks=params['logit_prop_tasks'],
-                    normalize_out = normalize_out)
-        elif "logit_prop_tasks" in params:
-            smiles, Y_logit = mu.load_smiles_and_data_df(params['data_file'], MAX_LEN,
-                    logit_tasks=params['logit_prop_tasks'], normalize_out=normalize_out)
-        elif "reg_prop_tasks" in params:
-            smiles, Y_reg = mu.load_smiles_and_data_df(params['data_file'], MAX_LEN,
-                    reg_tasks=params['reg_prop_tasks'], normalize_out=normalize_out)
-        else:
-            raise ValueError("please sepcify logit and/or reg tasks")
-
-    ## Load data if no properties
-    else:
-        smiles = mu.load_smiles_and_data_df(params['data_file'], MAX_LEN)
-
-    # print("NUMBER OF SAMPLES L<120: ", len(smiles))
-
-    if 'limit_data' in params.keys():
-        sample_idx = np.random.choice(np.arange(len(smiles)), params['limit_data'], replace=False)
-        smiles=list(np.array(smiles)[sample_idx])
-        if params['do_prop_pred'] and ('data_file' in params):
-            if "reg_prop_tasks" in params:
-                Y_reg =  Y_reg[sample_idx]
-            if "logit_prop_tasks" in params:
-                Y_logit =  Y_logit[sample_idx]
-
-    print('Training set size is', len(smiles))
-    print('first smiles: \"', smiles[0], '\"')
-    print('total chars:', NCHARS)
-
-    print('Vectorization...')
-    X = mu.smiles_to_hot(smiles, MAX_LEN, params[
-                             'PADDING'], CHAR_INDICES, NCHARS)
-
-    print('Total Data size', X.shape[0])
-    if np.shape(X)[0] % params['batch_size'] != 0:
-        X = X[:np.shape(X)[0] // params['batch_size']
-              * params['batch_size']]
-        if params['do_prop_pred']:
-            if "reg_prop_tasks" in params:
-                Y_reg = Y_reg[:np.shape(Y_reg)[0] // params['batch_size']
-                      * params['batch_size']]
-            if "logit_prop_tasks" in params:
-                Y_logit = Y_logit[:np.shape(Y_logit)[0] // params['batch_size']
-                      * params['batch_size']]
-
-    np.random.seed(params['RAND_SEED'])
-    rand_idx = np.arange(np.shape(X)[0])
-    np.random.shuffle(rand_idx)
-
-    TRAIN_FRAC = 1 - params['val_split']
-    num_train = int(X.shape[0] * TRAIN_FRAC)
-
-    if num_train % params['batch_size'] != 0:
-        num_train = num_train // params['batch_size'] * \
-            params['batch_size']
-
-    train_idx, test_idx = rand_idx[: int(num_train)], rand_idx[int(num_train):]
-
-    if 'test_idx_file' in params.keys():
-        np.save(params['test_idx_file'], test_idx)
-
-    X_train, X_test = X[train_idx], X[test_idx]
-    print('shape of input vector : {}', np.shape(X_train))
-    print('Training set size is {}, after filtering to max length of {}'.format(
-        np.shape(X_train), MAX_LEN))
-
-    if params['do_prop_pred']:
-        # !# add Y_train and Y_test here
-        Y_train = []
-        Y_test = []
-        if "reg_prop_tasks" in params:
-            Y_reg_train, Y_reg_test = Y_reg[train_idx], Y_reg[test_idx]
-            Y_train.append(Y_reg_train)
-            Y_test.append(Y_reg_test)
-        if "logit_prop_tasks" in params:
-            Y_logit_train, Y_logit_test = Y_logit[train_idx], Y_logit[test_idx]
-            Y_train.append(Y_logit_train)
-            Y_test.append(Y_logit_test)
-
-        return X_train, X_test, Y_train, Y_test
-
-    else:
-        return X_train, X_test
-
 
 def train_decoder(params):
     start_time = time.time()
@@ -321,21 +179,6 @@ def train_decoder(params):
     print('**FINISHED**')
 
     return
-
-def sigmoid_schedule(time_step, slope=1.0, start=None, weight_orig=None):
-    """
-    Sigmoid annealing.
-
-    :param time_step: epoch number
-    :param slope: slope of the sigmoid
-    :param start: starting point of the annealing
-    :param weight_orig: value of the weight at time_step=0
-    :return: sigmoid annealing weight
-    """
-    # Inverted float(time_step) and start wrt the original function
-    # The function should be decreasing with the time_step for weight annealing
-    # return weight_orig * float(1 / (1.0 + np.exp(slope * (float(time_step) - start))))
-    return weight_orig * float(1 / (1.0 + np.exp(slope * (start - float(time_step)))))
 
 
 if __name__ == "__main__":
